+++ src/app/forms/UserProfileForm.tsx
@@ -3,7 +3,7 @@ import {
   UserProfile,
   UserProfileMutation,
   userProfileMutationType,
-} from "../../api/services/auth/types";
+} from "../../api/services/user/types";
 import { useZodForm } from "../../lib/zod/useZodForm";
 import { TextField } from "../controls/TextField";
 import { getEnumName } from "../../lib/getEnumValue";
