+++ src/app/layout/Menu.tsx
@@ -2,7 +2,7 @@ import { Divider, Typography } from "@mui/material";
 import { router } from "../router";
 import { Auth } from "../components/Auth";
 import { RouteList } from "../components/RouteList";
-import { UserAccessLevel } from "../../api/services/auth/types";
+import { UserAccessLevel } from "../../api/services/user/types";
 
 const publicRoutes = [router.item, router.monster, router.map];
 const protectedRoutes = Object.values(router.admin.children);
