+++ src/app/pages/LoginPage.tsx
@@ -8,7 +8,7 @@ import { Link } from "../components/Link";
 import { CenteredContent } from "../components/CenteredContent";
 import { Header } from "../layout/Header";
 import { UserLoginForm } from "../forms/UserLoginForm";
-import { LoginPayload } from "../../api/services/auth/types";
+import { LoginPayload } from "../../api/services/user/types";
 
 export default function LoginPage() {
   const { destination } = useRouteParams(router.user().login);
